+++ CHANGELOG.md
@@ -73,6 +73,7 @@
 * Support writes past the end of custom SaveFile streams
 * Only edit input editor analog values after double-click
 * Fix Vulkan bug where vkCreateInstance would not be hooked
+* Avoid duplicate pthread ids
 
 ## [1.4.6] - 2024-07-05
 ### Added